[LCR-63] refactor: image 도메인 생성
