+++ src/main/java/com/lgcns/domain/image/controller/ImageController.java
@@ -1,7 +1,8 @@
-package com.lgcns.infra.s3;
+package com.lgcns.domain.image.controller;
 
-import com.lgcns.infra.s3.dto.request.FileNameRequest;
-import com.lgcns.infra.s3.dto.response.PreSignedUrlResponse;
+import com.lgcns.domain.image.dto.request.FileNameRequest;
+import com.lgcns.domain.image.dto.response.PreSignedUrlResponse;
+import com.lgcns.domain.image.service.ImageService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -15,15 +16,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/s3")
-public class S3Controller {
+public class ImageController {
 
-    private final S3Service s3Service;
+    private final ImageService imageService;
 
     @PostMapping("/presigned-url")
     @Operation(
             summary = "S3 preSigned url 생성",
             description = "파일 이름을 입력받아 S3 preSigned URL을 생성합니다.<br>" + "파일 이름에 파일 확장자는 필수입니다.")
     public PreSignedUrlResponse preSignedUrlCreate(@RequestBody @Valid FileNameRequest request) {
-        return s3Service.createPreSignedUrl("popi", request.fileName());
+        return imageService.createPreSignedUrl("popi", request.fileName());
     }
 }
